requirements.txt
discord.py
schedule
jsons